refactor(domain): 로또, 추첨번호 도메인 객체의 리팩토링
