+++ src/main/java/lotto/hexagon/application/RewardService.java
@@ -5,28 +5,18 @@ import java.util.Optional;
 import lotto.hexagon.domain.Award;
 import lotto.hexagon.domain.Drawn;
 import lotto.hexagon.domain.Lottos;
-import lotto.hexagon.domain.Match;
 import lotto.hexagon.domain.Prize;
 
 public class RewardService {
 
     public Award determine(Drawn drawn, Lottos lottos) {
-        List<Match> matches = getMatches(drawn, lottos);
-
-        List<Prize> prizes = findPrize(matches);
-
+        List<Prize> prizes = getEvaluatedPrizes(drawn, lottos);
         return getAward(prizes);
     }
 
-    private List<Match> getMatches(Drawn drawn, Lottos lottos) {
+    private List<Prize> getEvaluatedPrizes(Drawn drawn, Lottos lottos) {
         return lottos.stream()
-                .map(drawn::check)
-                .toList();
-    }
-
-    private List<Prize> findPrize(List<Match> matches) {
-        return matches.stream()
-                .map(Prize::findBy)
+                .map(lotto -> lotto.evaluate(drawn))
                 .flatMap(Optional::stream)
                 .toList();
     }
