+++ src/main/java/lotto/hexagon/domain/Drawn.java
@@ -1,12 +1,4 @@
 package lotto.hexagon.domain;
 
 public record Drawn(Numbers lucky, Number bonus) {
-
-    public Match check(Lotto lotto) {
-        Numbers numbers = lotto.getNumbers();
-        int luckyCount = lucky.getMatchCount(numbers);
-        boolean bonusMatch = numbers.has(bonus);
-
-        return Match.of(luckyCount, bonusMatch);
-    }
 }
