+++ src/main/java/lotto/hexagon/domain/Lotto.java
@@ -1,5 +1,7 @@
 package lotto.hexagon.domain;
 
+import java.util.Optional;
+
 public class Lotto {
     public static final Money PRICE = new Money(1000);
 
@@ -11,5 +13,20 @@ public class Lotto {
 
     public Numbers getNumbers() {
         return numbers;
+    }
+
+    public Optional<Prize> evaluate(Drawn drawn) {
+        Match match = check(drawn);
+        return Prize.findBy(match);
+    }
+
+    public Match check(Drawn drawn) {
+        Numbers lucky = drawn.lucky();
+        Number bonus = drawn.bonus();
+
+        int luckyCount = lucky.getMatchCount(numbers);
+        boolean bonusMatch = numbers.has(bonus);
+
+        return Match.of(luckyCount, bonusMatch);
     }
 }
